All tests are currently passing.
However, it's false.  There are still some oauth2 functions sitting around that need to be cleaned up.

Also, tests passing after the tests were modified is kind of suspicious.  That's because with Python 3 it was returning back HTTPretty.last_request.body as a byte string.  Using the decode function on them corrected this, however... I still haven't tested the updates against the actual Tumblr API, only the HTTPretty fake responses.  So I don't have a lot of faith in the tests showing passed currently.